Move back functionality to home. Need to pass params to hero
